Clean up akka-file-mailbox for OSGi. #see 2497
