+++ akka-durable-mailboxes/akka-file-mailbox/src/main/scala/akka/actor/mailbox/filebased/filequeue/BrokenItemException.scala
@@ -15,18 +15,8 @@
  * limitations under the License.
  */
 
-package akka.actor.mailbox.filequeue
+package akka.actor.mailbox.filebased.filequeue
 
-import java.util.concurrent.atomic.AtomicLong
+import java.io.IOException
 
-class Counter {
-  private val value = new AtomicLong(0)
-
-  def apply() = value.get
-  def set(n: Long) = value.set(n)
-  def incr() = value.addAndGet(1)
-  def incr(n: Long) = value.addAndGet(n)
-  def decr() = value.addAndGet(-1)
-  def decr(n: Long) = value.addAndGet(-n)
-  override def toString = value.get.toString
-}
+case class BrokenItemException(lastValidPosition: Long, cause: Throwable) extends IOException(cause)
